feature: split code on modules
