+++ constants/data.js
@@ -1,0 +1,6 @@
+export const MAX_COMMENTS_COUNT = 30;
+export const MAX_PHOTOS_COUNT = 25;
+export const commenterNames = ["Софья", "Ева", "Анна", "Мария", "Виктория", "Александр", "Михаил", "Артем", "Матвей", "Максим"];
+export const commentMessages = ["Всё отлично!", "В целом всё неплохо. Но не всё.", "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.", "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.", "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.", "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"];
+
+export const photoDescriptions = ["Запечатленная картина этого заката словно мечта, ожившая на фото.", "Каждый снимок дарит возможность вернуть в память миг, который невозможно забыть.", "На фотографии улыбка ребенка излучает счастье, способное согреть любое сердце.", "Старыe фото напоминают о прошедших временах, когда жизнь была другой.", "В каждом кадре скрыта история, достойная множества слов и эмоций. "];
